Address Aaron's PR feedback
* Moved the test under Loader/classloader/StaticVirtualMethods
* Added banner
* Added descriptive comment explaining the purpose of the test
