+++ src/tests/Loader/classloader/StaticVirtualMethods/RegressionTests/GitHub_74792.csproj
